+++ eventlog-live/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `verbosityInfo` and `verbosityDebug`.
 - **BREAKING**: Add logs to `runWithEventlogSocket` connection loop.
 - **BREAKING**: Downgrade severity of all errors logged during eventlog analyses to warnings.
+- **BREAKING**: Move attributes, metrics, and spans into their own modules.
 
 ### 0.1.0.1